fix: 위시리스트에 없는 상품은 주문 불가능하던 문제 수정

+++ src/main/java/gift/service/OrderService.java
@@ -51,8 +51,9 @@ public class OrderService {
         Order saved = orderRepository.save(order);
 
         Wish wish = wishRepository.findByMemberIdAndProductId(memberId, option.getProduct().getId());
-        wishRepository.delete(wish);
-
+        if  (wish != null) {
+            wishRepository.delete(wish);
+        }
         return saved;
     }
